+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added generation of GraphQL schema's Python representation.
 - Fixed annotations for lists.
+- Fixed support of custom operation types names.
 
 
 ## 0.4.0 (2023-03-20)